Setup to allow validation of nano vs cmg

+++ TTHAnalysis/cfg/validate_multilep.sh
@@ -1,3 +1,4 @@
+#!/bin/bash
 RUN=/bin/false;
 if [[ "$1" == "-run" ]]; then RUN=/bin/true; shift; fi;
 
@@ -10,6 +11,10 @@ function do_run {
     CFG=$1; shift
     name=$1; [[ "$name" == "" ]] && return; shift;
     echo "Will run as $name";
+    if test -f $name; then
+        echo "ERROR: $name is a file.";
+        return;
+    fi
     rm -r $name 2> /dev/null
     echo "heppy $name $CFG -p 0 -o nofetch $*"
     heppy $name $CFG -p 0 -o nofetch $*
@@ -17,6 +22,23 @@ function do_run {
     echo "Run done. press enter to continue (ctrl-c to break)";
     read DUMMY;
 }
+
+function do_nano {
+    CFG=$1; shift
+    name=$1; [[ "$name" == "" ]] && return; shift;
+    echo "Will run as $name";
+    if test -f $name; then
+        echo "ERROR: $name is a file.";
+        return;
+    fi
+    rm -r $name 2> /dev/null
+    echo "nanopy.py $name $CFG $*"
+    nanopy.py $name $CFG $*
+    if ls -1 $name/ | grep -q _Chunk0; then (cd $name; rm *_Chunk*/cmsswPreProcessing.root 2> /dev/null; haddChunks.py -n -c .); fi; 
+    echo "Run done. press enter to continue (ctrl-c to break)";
+    read DUMMY;
+}
+
 function do_friends_heppy {
     pushd ../macros
     SRC=$1; shift;
@@ -32,61 +54,59 @@ function do_friends_heppy {
             ;;
     esac;
     popd
+    echo "Run done. press enter to continue (ctrl-c to break)";
+    read DUMMY;
 }
 function do_plot {
     PROC=$1; PROCR=$2; LABEL=$3; RVER=$4; shift; shift; shift; shift;
     if [[ "${PROCR}" == "" ]]; then return; fi;
-    if test \! -d ${DIR}/${PROC} && test \! -d ${DIR}/New; then echo "Did not find ${PROC} or New in ${DIR}"; exit 1; fi
-    if [[ "${RVER}" == "" ]]; then RVER=94X; fi;
-    if [[ "${PROCR}" == "Ref" ]]; then
-        test -d ${DIR}/Ref && echo "Using existing Ref"
-    elif [[ "${LABEL}" != "MANUAL" ]]; then 
-        test -L ${DIR}/Ref && rm ${DIR}/Ref    
-        test -L ${DIR}/New && rm ${DIR}/New    
-        if test -d ~/Reference_${RVER}_${PROCR}${LABEL}; then
-             ln -sd ~/Reference_${RVER}_${PROCR}${LABEL} ${DIR}/Ref;
-        else
-             ln -sd $PWD/Reference_${RVER}_${PROCR}${LABEL} ${DIR}/Ref;
-        fi
-        ln -sd ${DIR}/${PROC} ${DIR}/New;
+    if test  -d ${DIR}/${PROC} || test -d ${DIR}/New; then 
+        NANOAOD=false; TESTARG="-d"; LNARG="-sd"; POST="";
+    elif test  -f ${DIR}/${PROC}.root || test -f ${DIR}/New.root; then 
+        NANOAOD=true; TESTARG="-f"; LNARG="-s"; POST=".root";
     else
-        test -L ${DIR}/Ref && rm ${DIR}/Ref    
-        test -L ${DIR}/New && rm ${DIR}/New    
-        if test -d ${DIR}/${PROCR}; then
-            ln -sd ${DIR}/${PROCR} ${DIR}/Ref
-        elif test -d ${PROCR}; then
-            ln -sd ${PROCR}  ${DIR}/Ref
-            PROCR=$(basename ${PROCR})
-        fi;
-        ln -sd ${DIR}/${PROC} ${DIR}/New;
+        echo "Did not find ${PROC} or New in ${DIR}"; exit 1; 
     fi
-    test -d ${DIR}/New || ln -sd ${DIR}/${PROC} ${DIR}/New;
+    if [[ "${RVER}" == "" ]]; then RVER=94X; fi;
+    if [[ "${KEEP_REF}" == "1" ]]; then
+        if test \! -d ${DIR}/Ref && test \! -f ${DIR}/Ref.root; then
+            echo "KEEP_REF set to 1 but I can't find a reference"; exit 1;
+        fi
+    elif [[ "${LABEL}" != "MANUAL" ]]; then 
+        for X in Ref Ref.root New New.root; do test -L ${DIR}/$X && rm ${DIR}/$X; done
+        if test ${TESTARG} ~/Reference_${RVER}_${PROCR}${LABEL}${POST}; then
+             ln $LNARG ~/Reference_${RVER}_${PROCR}${LABEL}${POST} ${DIR}/Ref${POST};
+        elif test ${TESTARG} $PWD/Reference_${RVER}_${PROCR}${LABEL}${POST}; then
+             ln $LNARG $PWD/Reference_${RVER}_${PROCR}${LABEL}${POST} ${DIR}/Ref${POST};
+        else
+             echo "No idea where to take the reference from"; exit 1; 
+        fi
+    else
+        for X in Ref Ref.root New New.root; do test -L ${DIR}/$X && rm ${DIR}/$X; done
+        if test ${TESTARG} ${DIR}/${PROCR}${POST}; then
+            ln $LNARG ${DIR}/${PROCR}${POST} ${DIR}/Ref${POST}
+        elif test ${TESTARG} ${PROCR}${POST}; then
+            ln $LNARG ${PROCR}${POST}  ${DIR}/Ref${POST}
+            PROCR=$(basename ${PROCR} ${POST})
+        fi;
+    fi
+    test ${TESTARG} ${DIR}/New${POST} || ln $LNARG ${DIR}/${PROC}${POST} ${DIR}/New${POST};
     OUTNAME=${WHAT}-${PROCR}${LABEL}
     ( cd ../python/plotter;
-      # ---- MCA ---
-      MCA=susy-multilepton/validation_mca.txt
-      if echo $PROC | grep -q Run201[2567]; then MCA=susy-multilepton/validation-data_mca.txt; fi;
-      # ---- CUT FILE ---
-      CUTS=susy-multilepton/validation.txt;
-      if [ -f susy-multilepton/validation-${PROC}.txt ]; then 
-        CUTS=susy-multilepton/validation-${PROC}.txt
-      elif echo $PROC | grep -q Run201[67]; then
-        if echo $PROC | grep -q Single; then
-             CUTS=susy-multilepton/validation-data-single.txt
-        elif echo $PROC | grep -q 'MET\|2017'; then
-             CUTS=susy-multilepton/validation.txt
-        else
-             CUTS=susy-multilepton/validation-data.txt
-        fi;
-      elif echo "X$LABEL" | grep -q [Rr]ecl; then
-        CUTS=ttH-multilepton/validation-recl.txt
-      fi
-      WA=1; if echo $WHAT | grep -q Presc; then WA=prescaleFromSkim; fi;
-      echo "python mcPlots.py -f --s2v --tree treeProducerSusyMultilepton  -P ${DIR} $MCA $CUTS ${CUTS/.txt/_plots.txt} \
-              --pdir plots/104X/validation/${OUTNAME}  -u -e --WA $WA $* \
+      # ---- MCA, CUT & PLOT FILE ---
+      [[ "${MCA}" == "" ]] && MCA=susy-multilepton/validation_mca.txt
+      [[ "${CUTS}" == "" ]] && CUTS=susy-multilepton/validation.txt
+      [[ "${PLOTS}" == "" ]] && PLOTS=${CUTS/.txt/_plots.txt}
+      if $NANOAOD; then
+          TREE=NanoAOD
+      else
+          [[ "${TREE}" == "" ]] && TREE=treeProducerSusyMultilepton
+      fi;
+      echo "python mcPlots.py -f --s2v --tree $TREE  -P ${DIR} $MCA $CUTS ${PLOTS} \
+              --pdir plots/104X/validation/${OUTNAME}  -u -e $* \
               --plotmode=nostack --showRatio --maxRatioRange 0.65 1.35 --flagDifferences"
-      python mcPlots.py -f --s2v --tree treeProducerSusyMultilepton  -P ${DIR} $MCA $CUTS ${CUTS/.txt/_plots.txt} \
-              --pdir plots/104X/validation/${OUTNAME}  -u -e --WA $WA $* \
+      python mcPlots.py -f --s2v --tree $TREE  -P ${DIR} $MCA $CUTS ${PLOTS} \
+              --pdir plots/104X/validation/${OUTNAME}  -u -e $* \
               --plotmode=nostack --showRatio --maxRatioRange 0.65 1.35 --flagDifferences
     );
 }
@@ -99,6 +119,7 @@ function do_size {
 case $WHAT in
     ttHData)
         $RUN && do_run run_ttH_cfg.py $DIR -o test=94X-Data  -N 10000 -o runData;
+        CUTS=susy-multilepton/validation-data.txt
         do_plot DoubleMuon_Run2017C DoubleMuon_Run2017C
         do_plot DoubleEG_Run2017E DoubleEG_Run2017E
         ;;
@@ -106,17 +127,76 @@ case $WHAT in
         $RUN && do_run run_ttH_cfg.py $DIR -o test=94X-MC -o sample=TTLep -N 2000;
         do_plot TTLep_pow TTLep_pow
         ;;
-    ttHMC_Recl)
-        if test -d ${DIR/_Recl/}; then
+    ttHMC-recl)
+        if test -d ${DIR/-recl/}; then
             if $RUN; then
                 test -d $DIR || mkdir $DIR || exit 1;
-                test -d $DIR/New || cp -av ${DIR/_Recl/}/{Ref,New,TTLep_pow} $DIR || exit 1;
-                test -d $DIR/1_recleaner_v0 || cp -rv ~/Reference_94X_TTLep_pow_friends/[0-9]_* $DIR || exit 1;
+                test -d $DIR/New || cp -a ${DIR/-recl/}/{Ref,New,TTLep_pow} $DIR || exit 1;
+                test -d $DIR/1_recleaner_v0 || cp -r ~/Reference_94X_TTLep_pow_friends/[0-9]_* $DIR || exit 1;
                 do_friends_heppy $DIR ttHMC New;
             fi;
-            do_plot TTLep_pow Ref -recl 94X --Fs {P}/1_recleaner_v0 --Fs {P}/2_eventVars_v0  -L ttH-multilepton/functionsTTH.cc --mcc ttH-multilepton/lepchoice-ttH-FO.txt
+            PLOTS=ttH-multilepton/validation/recl_plots.txt
+            KEEP_REF=1
+            do_plot TTLep_pow TTLep_pow "" "" --Fs {P}/1_recleaner_v0 --Fs {P}/2_eventVars_v0 -L ttH-multilepton/functionsTTH.cc --mcc ttH-multilepton/lepchoice-ttH-FO.txt
         else
-            echo "first run ttHMC to create $DIR";
+            echo "first run ttHMC";
+        fi
+        ;;
+    nano_ttHMC)
+        $RUN && do_nano run_ttH_fromNanoAOD_cfg.py $DIR -o test=94X-MC
+        PLOTS=ttH-multilepton/validation/nanoaod_plots.txt
+        do_plot TTLep_pow TTLep_pow
+        ;;
+    nano_ttHMC-vs_cmg)
+        BASE=${DIR/nano_ttHMC-vs_cmg/}
+        if test -d ${BASE}ttHMC && test -d ${BASE}nano_ttHMC; then
+            test -d $DIR || mkdir $DIR || exit 1;
+            test -L $DIR/New.root || ln -s  ${BASE}nano_ttHMC/New.root $DIR/New.root 
+            test -L $DIR/Ref      || ln -sd ${BASE}ttHMC/New           $DIR/Ref
+            MCA=ttH-multilepton/validation/nanoaod-vs-cmg_mca.txt
+            PLOTS=ttH-multilepton/validation/nanoaod_plots.txt
+            KEEP_REF=1
+            do_plot TTLep_pow TTLep_pow
+        else
+            echo "first run ttHMC and nano_ttHMC";
+        fi
+        ;;
+    nano_ttHMC-recl)
+        if test -d ${DIR/-recl/}; then
+            if $RUN; then
+                test -d $DIR || mkdir $DIR || exit 1;
+                test -d $DIR/New.root || cp -a ${DIR/-recl/}/{Ref.root,New.root} $DIR || exit 1;
+                test -d $DIR/1_recleaner_v0 || cp -r ~/Reference_94X_TTLep_pow_nanoaod_friends/[0-9]_* $DIR || exit 1;
+                MODS=CMGTools.TTHAnalysis.tools.nanoAOD.ttH_modules
+                nano_postproc.py -I $MODS recleaner_step1,recleaner_step2_mc --friend ${DIR}/1_recleaner_v0 ${DIR}/New.root
+                nano_postproc.py -I $MODS eventVars                          --friend ${DIR}/2_eventVars_v0 ${DIR}/New.root,${DIR}/1_recleaner_v0/New_Friend.root
+                echo "Run done. press enter to continue (ctrl-c to break)";
+                read DUMMY;
+            fi;
+            PLOTS=ttH-multilepton/validation/nanoaod-recl_plots.txt
+            KEEP_REF=1
+            do_plot TTLep_pow TTLep_pow "" "" --Fs {P}/1_recleaner_v0 --Fs {P}/2_eventVars_v0 -L ttH-multilepton/functionsTTH.cc --mcc ttH-multilepton/lepchoice-ttH-FO.txt
+        else
+            echo "first run nano_ttHMC";
+        fi
+        ;;
+    nano_ttHMC-recl-vs_cmg)
+        BASE=${DIR/nano_ttHMC-recl-vs_cmg/}
+        if test -d ${BASE}ttHMC-recl && test -d ${BASE}nano_ttHMC-recl; then
+            test -d $DIR || mkdir $DIR || exit 1;
+            test -L $DIR/New.root || ln -s  ${BASE}nano_ttHMC-recl/New.root $DIR/New.root 
+            test -L $DIR/Ref      || ln -sd ${BASE}ttHMC-recl/New           $DIR/Ref
+            for F in 1_recleaner_v0 2_eventVars_v0; do
+                test -d $DIR/$F || mkdir $DIR/$F || exit 1;
+                test -L $DIR/$F/New_Friend.root      || ln -s ${BASE}nano_ttHMC-recl/$F/New_Friend.root  $DIR/$F/
+                test -L $DIR/$F/evVarFriend_Ref.root || ln -s ${BASE}ttHMC-recl/$F/evVarFriend_Ref.root $DIR/$F/
+            done
+            MCA=ttH-multilepton/validation/nanoaod-vs-cmg_mca.txt
+            PLOTS=ttH-multilepton/validation/nanoaod-recl_plots.txt
+            KEEP_REF=1
+            do_plot TTLep_pow TTLep_pow "" "" --Fs {P}/1_recleaner_v0 --Fs {P}/2_eventVars_v0 -L ttH-multilepton/functionsTTH.cc --mcc ttH-multilepton/lepchoice-ttH-FO.txt
+        else
+            echo "first run ttHMC-recl and nano_ttHMC-recl";
         fi
         ;;
     ttHSpeed)
@@ -130,6 +210,7 @@ case $WHAT in
         ;;
     ttHDataSize)
         $RUN && do_run run_ttH_cfg.py $DIR -o test=94X-Data  -N 100000 -o runData -o fast;
+        CUTS=susy-multilepton/validation-data.txt
         do_size DoubleMuon_Run2017C 
         do_size DoubleEG_Run2017E
         do_plot DoubleMuon_Run2017C DoubleMuon_Run2017C _big
@@ -137,24 +218,29 @@ case $WHAT in
         ;;
     ttHDataPresc)
         $RUN && do_run run_ttH_cfg.py $DIR -o test=94X-Data  -N 100000 -o runData -o fast -o prescaleskim;
-        do_plot DoubleMuon_Run2017C DoubleMuon_Run2017C _big
-        do_plot DoubleEG_Run2017E DoubleEG_Run2017E _big
+        CUTS=susy-multilepton/validation-data.txt
+        do_plot DoubleMuon_Run2017C DoubleMuon_Run2017C _big 94X --WA prescaleFromSkim
+        do_plot DoubleEG_Run2017E DoubleEG_Run2017E _big 94X --WA prescaleFromSkim
         ;;
     ttHData80X)
         $RUN && do_run run_ttH_cfg.py $DIR -o test=80X-Data  -N 10000 -o runData -o run80X;
+        CUTS=susy-multilepton/validation-data.txt
         do_plot DoubleMuon_Run2016H_run283885 DoubleMuon_Run2016H_run283885 "" 80X
         do_plot DoubleEG_Run2016H_run283885 DoubleEG_Run2016H_run283885 "" 80X
         ;;
     ttHMC80X)
         $RUN && do_run run_ttH_cfg.py $DIR -o test=80X-MC -o sample=TTLep  -o run80X -N 2000;
+        CUTS=susy-multilepton/validation-data.txt
         do_plot TTLep_pow TTLep_pow "" 80X
         ;;
     SOSData)
         $RUN && do_run run_susyMultilepton_cfg.py $DIR -o test=94X-Data  -N 40000 -o runData -o sample=DoubleMuon  -o analysis=SOS;
+        CUTS=susy-multilepton/validation-data.txt
         do_plot DoubleMuon_Run2017C_run299649 DoubleMuon_Run2017C_run299649 _SOS
         ;;
     SOSData80X)
         $RUN && do_run run_susyMultilepton_cfg.py $DIR -o test=80X-Data  -N 100000 -o runData -o sample=MET  -o analysis=SOS;
+        CUTS=susy-multilepton/validation.txt
         do_plot MET_Run2016H_run283885 MET_Run2016H_run283885 _SOS 80X
         ;;
     SOSMC)
